requirements.txt
aiohttp==3.7.4.post0
anaconda-navigator==1.8.7
anaconda-project==0.8.2
astropy==3.0.2
atomicwrites==1.4.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bokeh==0.12.16
boto==2.48.0
bottleneck==1.2.1
brotlipy==0.7.0
bs4==0.0.1
cached-property==1.5.2
colorama==0.3.9
conda-build==3.11.0
conda-package-handling==1.7.0
contextlib2==0.5.5
cython==0.28.2
cytoolz==0.9.0.1
distributed==1.21.8
dnspython==1.16.0
fastcache==1.0.2
gevent==1.3.0
gmaps==0.9.0
gmpy2==2.0.8
google-auth-oauthlib==0.4.4
html5lib==1.0.1
importlib==1.0.4
imutils==0.5.3
jupyter==1.0.0
jupyterlab==2.2.4
jupyterlab-launcher==0.10.5
keras==2.2.4
keyring==22.3.0
lxml==4.5.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.0.0
msgpack-python==0.5.6
navigator-updater==0.2.1
nltk==3.3
nose==1.3.7
numba==0.38.0
numpydoc==0.8.0
olefile==0.45.1
openpyxl==2.5.3
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pdf2image==1.14.0
pep8==1.7.1
pip-chill==1.0.1
ply==3.11
prettytable==2.1.0
pycodestyle==2.4.0
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
pylint==1.8.4
pymongo==3.11.4
pyodbc==4.0.23
pyopenssl==19.0.0
pysocks==1.7.1
pytest-astropy==0.3.0
python-decouple==3.4
python-dotenv==0.19.0
python-whois==0.7.3
qtawesome==0.4.4
regex==2020.10.11
requests-file==1.5.1
rope==0.10.7
scikit-image==0.17.2
seaborn==0.10.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
sklearn==0.0
sortedcollections==0.6.1
spotlight==0.1.5
spyder==3.2.8
spyder-kernels==1.4.0
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.9
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.12.0
tqdm==4.49.0
typed-ast==1.4.2
typing==3.6.4
unicodecsv==0.14.1
virtualenv==20.4.3
xlrd==1.1.0
xlsxwriter==1.0.4
xlwt==1.3.0
